[2024-05-24] controller 정리
