+++ src/main/java/com/ohgiraffers/hitechautoworks/auth/controller/UserController.java
@@ -31,13 +31,6 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-    @Autowired
-    private ResService resService;
-
-    @Autowired
-    private PartService partService;
-
-
     @ModelAttribute
     public void addUserToModel(Model model){
         AuthUserInfo authUserInfo = new AuthUserInfo();
@@ -133,74 +126,6 @@ public class UserController {
     }
 
 
-
-    @GetMapping("/user/common")
-    public void common(Model model) {
-        List<ResDTO> resList = resService.findAllres();
-        model.addAttribute("resList", resList);
-    }
-
-
-    @PostMapping("/user/res/ressearch")
-    public String resgo(Model model,@RequestParam String resCode, @RequestParam String resName){
-
-        if(resCode != "" && resName == ""){
-            int resCodeInt = Integer.parseInt(resCode);
-            ResDTO resList = resService.findUserRes(resCodeInt);
-            model.addAttribute("resList", resList);
-        } else if (resCode == "" && resName != ""){
-            List<ResDTO> resList = resService.findNameRes(resName);
-            model.addAttribute("resList", resList);
-        } else {
-            List<ResDTO> resList = resService.findAllres();
-            model.addAttribute("resList", resList);
-        }
-
-        return "user/common";
-    }
-
-    @GetMapping("/user/partAllCall")
-    public void partAllCall(Model model) {
-
-        List<PartDTO> partList = partService.selectAllPart();
-        model.addAttribute("partList", partList);
-    }
-
-    @PostMapping("/user/partAllCall")
-    public String partAllCall2(@RequestParam String partName, @RequestParam String partCode,
-                               Model model) {
-
-        if (partName == "" && partCode == "") {
-            List<PartDTO> partList = partService.selectAllPart();
-            model.addAttribute("partList", partList);
-            System.out.println("partList = " + partList);
-        } else if (partName == "") {
-            List<PartDTO> partList = partService.selectPartByCode(Integer.parseInt(partCode));
-            System.out.println("partList = " + partList);
-            model.addAttribute("partList", partList);
-        } else {
-            List<PartDTO> partList = partService.partSearchBtPartName(partName);
-            System.out.println("partList = " + partList);
-            model.addAttribute("partList", partList);
-        }
-
-        return "user/partAllCall";
-    }
-
-    @PostMapping("/user/registpart")
-    public String registpart(@RequestParam Map<String, String> parts, Model model){
-
-        int result = partService.addPart(parts);
-        if(result == 1) {
-            String partName = parts.get("partName");
-            model.addAttribute("result", result);
-            model.addAttribute("partName", partName);
-        }
-
-        return "redirect:/user/partAllCall";
-    }
-
-
     @PostMapping("/user/mypage/update")
     public String updateUser(@RequestParam Map<String, String> myprofile) {
 
@@ -212,80 +137,6 @@ public class UserController {
         userService.updateUser(myprofile);
 
         return "redirect:/user/mypage";
-    }
-
-    @GetMapping("/user/partAdd")
-    public void partAdd(Model model) {
-    }
-
-    @GetMapping("/user/testPage")
-    public void testpage(Model model, @RequestParam int resCode, HttpSession session) {
-
-        ResDTO res = resService.findUserRes(resCode / 123456);// 들어올때 resCode 123456 나눠줘야댐 (나중에 제대로 암호화 ㄱㄱ)
-        model.addAttribute("res", res);
-        String date = String.valueOf(res.getDate());
-        String repair = resService.findStatus(resCode/123456);
-        if (repair == null){
-            repair = "대기";
-        }
-        model.addAttribute("repair", repair);
-        model.addAttribute("sqldate", date.substring(0,19));
-//        List<ResCommentDTO> resCommentDTO = resService.findComment(resCode / 123456);
-//        model.addAttribute("resComment", resCommentDTO);
-
-        if (session.getAttribute("result") != null) {
-            model.addAttribute("result", session.getAttribute("result"));
-            session.removeAttribute("result");
-        }
-
-    }
-
-    @PostMapping("/user/registcomment")
-    public String testPage2(@RequestParam String comment, @RequestParam int resCode,Model model) {
-
-        int userCode = ((UserDTO) model.getAttribute("userDTO")).getUserCode();
-        resService.registcomment(comment,resCode,userCode);
-        return "redirect:/user/testPage?resCode=" + 123456 * resCode;
-    }
-
-    @PostMapping("/user/resUpdate")
-    public String resModify(@RequestParam int resCode ,@RequestParam String fixOption,@RequestParam String date,@RequestParam String extra, Model model,
-                            HttpSession session){
-
-        int result = resService.resModify(resCode,fixOption,date,extra);
-        if (result == 1){
-            session.setAttribute("result",result);
-        }
-        return "redirect:/user/testPage?resCode=" + 123456 * resCode;
-    }
-
-    @PostMapping("/user/resDelete")
-    public String resDelete(@RequestParam int resCode){
-        resService.resDelete(resCode);
-
-        return "user/testPage";
-    }
-
-
-    @GetMapping("/user/rescustomer")
-    public void resccustomer(Model model) {
-
-        int userCode = ((UserDTO) model.getAttribute("userDTO")).getUserCode();
-        List<ResDTO> resList = resService.findCustomerRes(userCode);
-        model.addAttribute("resList",resList);
-    }
-
-
-    @PostMapping("/user/reseditComment")
-    @ResponseBody
-    public int editComment(@RequestBody EditCommentDTO editCommentDTO){
-        int resReplyCode = editCommentDTO.getResReplyCode();
-        String editcomment = editCommentDTO.getStr();
-        int rescode = editCommentDTO.getRescode();
-
-        int result = resService.updateComment(resReplyCode, editcomment);
-
-        return result;
     }
 
 
@@ -312,28 +163,6 @@ public class UserController {
         return "redirect:/user/customermypage?customerUserCode=" + customerUserCode;
     }
 
-    @PostMapping("/user/deleteComment")
-    @ResponseBody
-    public int deleteComment(@RequestBody DeleteCommentDTO deleteCommentDTO){
-        int resReplyCode = deleteCommentDTO.getResReplyCode();
-        int rescode = deleteCommentDTO.getRescode();
-        int result = resService.deleteComment(resReplyCode);
-
-        return result;
-    }
-
-    @GetMapping("/user/res")
-    public String res() {
-
-        return "user/res";
-    }
-
-    @GetMapping("/user/selectRes")
-    public String selectRes() {
-
-        return "user/selectRes";
-    }
-
     @PostMapping("/user/res/resTime")
     @ResponseBody
     public Map<String, Object> checkResTime(@RequestBody Map<String,Date> date) {
@@ -353,40 +182,6 @@ public class UserController {
         return disabledTimes;
     }
 
-    @PostMapping("/user/res/Submit")
-    @ResponseBody
-    public String resSubmit(@RequestBody Map<String,Object> info,Model model){
-        String date = (String) info.get("date");
-        String time = (String) info.get("selectedRadioValue");
-        String dateTime = date+' '+time;
-        System.out.println("dateTime = " + dateTime);
-        for(String key : info.keySet()) {
-            String value = (String) info.get(key);
-            System.out.println(key + " : " + value);
-        }
-        String option = (String) info.get("message");
-        String resExtra = (String) info.get("resExtra");
-        int userCode = ((UserDTO) model.getAttribute("userDTO")).getUserCode();
-        resService.insertRes(userCode,option,dateTime,resExtra);
-
-        return "1";
-    }
-
-
-    // 작업중
-    @GetMapping("/user/resCar")
-    public String resCar() {
-
-        return "user/resCar";
-    }
-
-    @PostMapping("/user/res/carSubmit")
-    public String carSubmit(@RequestParam String inputMessage) {
-
-        System.out.println("inputMessage = " + inputMessage);
-
-        return "user/res";
-    }
 
     @PostMapping("/user/getCalendar")
     @ResponseBody
