+++ src/main/java/com/ohgiraffers/hitechautoworks/part/controller/PartController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
+import java.util.Map;
 
 @Controller
 public class PartController {
@@ -57,12 +58,47 @@ public class PartController {
         }
     }
 
-    @GetMapping("/employee/part/partAdd")
-    public void partAdd(Model model){
-        authUserInfo = new AuthUserInfo();
-        UserDTO userDTO = authUserInfo.getUserDTO();
-        String userName = userDTO.getUserName();
-        model.addAttribute("userName",userName);
+
+    @GetMapping("/user/partAllCall")
+    public void partAllCall(Model model) {
+
+        List<PartDTO> partList = partService.selectAllPart();
+        model.addAttribute("partList", partList);
+    }
+    @PostMapping("/user/partAllCall")
+    public String partAllCall2(@RequestParam String partName, @RequestParam String partCode,
+                               Model model) {
+
+        if (partName == "" && partCode == "") {
+            List<PartDTO> partList = partService.selectAllPart();
+            model.addAttribute("partList", partList);
+            System.out.println("partList = " + partList);
+        } else if (partName == "") {
+            List<PartDTO> partList = partService.selectPartByCode(Integer.parseInt(partCode));
+            System.out.println("partList = " + partList);
+            model.addAttribute("partList", partList);
+        } else {
+            List<PartDTO> partList = partService.partSearchBtPartName(partName);
+            System.out.println("partList = " + partList);
+            model.addAttribute("partList", partList);
+        }
+
+        return "user/partAllCall";
+    }
+    @PostMapping("/user/registpart")
+    public String registpart(@RequestParam Map<String, String> parts, Model model){
+
+        int result = partService.addPart(parts);
+        if(result == 1) {
+            String partName = parts.get("partName");
+            model.addAttribute("result", result);
+            model.addAttribute("partName", partName);
+        }
+
+        return "redirect:/user/partAllCall";
+    }
+    @GetMapping("/user/partAdd")
+    public void partAdd(Model model) {
     }
