+++ src/main/java/com/ohgiraffers/hitechautoworks/res/controller/ResController.java
@@ -24,6 +24,7 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Controller
@@ -207,38 +208,7 @@ public class ResController {
 
     }
 
-    @PostMapping("/customer/res/resDelete")
-    public String resDelete(@RequestParam int resCode ){
-     resService.resDelete(resCode);
 
-     return "customer/res/res";
-    }
-
-    @PostMapping("/user/res/editComment")
-    @ResponseBody
-    public int editComment(@RequestBody EditCommentDTO editCommentDTO){
-        int resReplyCode = editCommentDTO.getResReplyCode();
-        String editcomment = editCommentDTO.getStr();
-        int rescode = editCommentDTO.getRescode();
-
-        int result = resService.updateComment(resReplyCode, editcomment);
-
-//        return "redirect:/customer/res/resDetail?resCode=" + rescode;
-
-        return result;
-    }
-
-    @PostMapping("/customer/res/deleteComment")
-    @ResponseBody
-    public int deleteComment(@RequestBody DeleteCommentDTO deleteCommentDTO){
-        int resReplyCode = deleteCommentDTO.getResReplyCode();
-        int rescode = deleteCommentDTO.getRescode();
-        System.out.println("deleteCommentDTO111 = " + deleteCommentDTO);
-        int result = resService.deleteComment(resReplyCode);
-
-        return result;
-//        return "/customer/res/resDetail?resCode=" + rescode;
-    }
 
     @GetMapping("/customer/res/repair")
     public void resRepair(@RequestParam("resCode") int resCode,Model model) {
@@ -247,7 +217,146 @@ public class ResController {
         model.addAttribute("res", res);
     }
 
+    @GetMapping("/user/common")
+    public void common(Model model) {
+        List<ResDTO> resList = resService.findAllres();
+        model.addAttribute("resList", resList);
+    }
+    @PostMapping("/user/res/ressearch")
+    public String resgo(Model model,@RequestParam String resCode, @RequestParam String resName){
+
+        if(resCode != "" && resName == ""){
+            int resCodeInt = Integer.parseInt(resCode);
+            ResDTO resList = resService.findUserRes(resCodeInt);
+            model.addAttribute("resList", resList);
+        } else if (resCode == "" && resName != ""){
+            List<ResDTO> resList = resService.findNameRes(resName);
+            model.addAttribute("resList", resList);
+        } else {
+            List<ResDTO> resList = resService.findAllres();
+            model.addAttribute("resList", resList);
+        }
+
+        return "user/common";
+    }
+    @GetMapping("/user/testPage")
+    public void testpage(Model model, @RequestParam int resCode, HttpSession session) {
+
+        ResDTO res = resService.findUserRes(resCode / 123456);// 들어올때 resCode 123456 나눠줘야댐 (나중에 제대로 암호화 ㄱㄱ)
+        model.addAttribute("res", res);
+        String date = String.valueOf(res.getDate());
+        String repair = resService.findStatus(resCode/123456);
+        if (repair == null){
+            repair = "대기";
+        }
+        model.addAttribute("repair", repair);
+        model.addAttribute("sqldate", date.substring(0,19));
+//        List<ResCommentDTO> resCommentDTO = resService.findComment(resCode / 123456);
+//        model.addAttribute("resComment", resCommentDTO);
+
+        if (session.getAttribute("result") != null) {
+            model.addAttribute("result", session.getAttribute("result"));
+            session.removeAttribute("result");
+        }
+
+    }
+    @PostMapping("/user/registcomment")
+    public String testPage2(@RequestParam String comment, @RequestParam int resCode,Model model) {
+
+        int userCode = ((UserDTO) model.getAttribute("userDTO")).getUserCode();
+        resService.registcomment(comment,resCode,userCode);
+        return "redirect:/user/testPage?resCode=" + 123456 * resCode;
+    }
+
+    @PostMapping("/user/resUpdate")
+    public String resModify(@RequestParam int resCode ,@RequestParam String fixOption,@RequestParam String date,@RequestParam String extra, Model model,
+                            HttpSession session){
+
+        int result = resService.resModify(resCode,fixOption,date,extra);
+        if (result == 1){
+            session.setAttribute("result",result);
+        }
+        return "redirect:/user/testPage?resCode=" + 123456 * resCode;
+    }
+
+    @PostMapping("/user/resDelete")
+    public String resDelete(@RequestParam int resCode){
+        resService.resDelete(resCode);
+
+        return "user/testPage";
+    }
 
 
+    @GetMapping("/user/rescustomer")
+    public void resccustomer(Model model) {
+
+        int userCode = ((UserDTO) model.getAttribute("userDTO")).getUserCode();
+        List<ResDTO> resList = resService.findCustomerRes(userCode);
+        model.addAttribute("resList",resList);
+    }
+
+
+    @PostMapping("/user/reseditComment")
+    @ResponseBody
+    public int editComment(@RequestBody EditCommentDTO editCommentDTO){
+        int resReplyCode = editCommentDTO.getResReplyCode();
+        String editcomment = editCommentDTO.getStr();
+        int rescode = editCommentDTO.getRescode();
+
+        int result = resService.updateComment(resReplyCode, editcomment);
+
+        return result;
+    }
+    @PostMapping("/user/deleteComment")
+    @ResponseBody
+    public int deleteComment(@RequestBody DeleteCommentDTO deleteCommentDTO){
+        int resReplyCode = deleteCommentDTO.getResReplyCode();
+        int rescode = deleteCommentDTO.getRescode();
+        int result = resService.deleteComment(resReplyCode);
+
+        return result;
+    }
+    @GetMapping("/user/res")
+    public String res() {
+
+        return "user/res";
+    }
+
+    @GetMapping("/user/selectRes")
+    public String selectRes() {
+
+        return "user/selectRes";
+    }
+    @PostMapping("/user/res/Submit")
+    @ResponseBody
+    public String resSubmit(@RequestBody Map<String,Object> info, Model model){
+        String date = (String) info.get("date");
+        String time = (String) info.get("selectedRadioValue");
+        String dateTime = date+' '+time;
+        System.out.println("dateTime = " + dateTime);
+        for(String key : info.keySet()) {
+            String value = (String) info.get(key);
+            System.out.println(key + " : " + value);
+        }
+        String option = (String) info.get("message");
+        String resExtra = (String) info.get("resExtra");
+        int userCode = ((UserDTO) model.getAttribute("userDTO")).getUserCode();
+        resService.insertRes(userCode,option,dateTime,resExtra);
+
+        return "1";
+    }
+    @GetMapping("/user/resCar")
+    public String resCar() {
+
+        return "user/resCar";
+    }
+
+    @PostMapping("/user/res/carSubmit")
+    public String carSubmit(@RequestParam String inputMessage) {
+
+        System.out.println("inputMessage = " + inputMessage);
+
+        return "user/res";
+    }
 
 }
