+++ src/main/resources/templates/user/testPage.html
@@ -135,10 +135,16 @@
                                     <input type="text" name="resCode" class="form-control" th:value="${res.code}" readonly>
                                 </div>
                             </div>
+                            <div class="row mb-3 mt-4">
+                                <label class="col-sm-3 col-form-label text-end">예약자</label>
+                                <div class="col-sm-6">
+                                    <input type="text" name="resCode" class="form-control" th:value="${res.userDTO.userName}" readonly>
+                                </div>
+                            </div>
                             <div class="row mb-3">
                                 <label class="col-sm-3 col-form-label text-end">수리옵션</label>
                                 <div class="col-sm-6">
-                                    <input type="text" name="fixOption" class="form-control" th:value="${res.option}">
+                                    <input type="text" name="fixOption" class="form-control" th:value="${res.option}" readonly>
                                 </div>
                             </div>
                             <div class="row mb-3">
@@ -153,12 +159,12 @@
                                     <input type="text" name="extra" class="form-control" th:value="${res.extra}">
                                 </div>
                             </div>
-                                <div class="row mb-3">
-                                    <label class="col-sm-3 col-form-label text-end">진행상황</label>
-                                    <div class="col-sm-6">
-                                        <input type="text" name="status" class="form-control" th:value="${repair}" readonly>
-                                    </div>
+                            <div class="row mb-3">
+                                <label class="col-sm-3 col-form-label text-end">진행상황</label>
+                                <div class="col-sm-6">
+                                    <input type="text" name="status" class="form-control" th:value="${repair}" readonly>
                                 </div>
+                            </div>
                             <div class="row">
                                 <div class="col text-end">
